Export COMPRESSION_VERSION from package
This constant provides the version of the compression algorithm. Other
npm packages can use this to verify that data is being compressed and
decompressed using the same algorithm. For example, a data store may
keep the compression version alongside the compressed data. Then, when
decompressing, the store can verify that the current compression version
matches the stored compression version or throw an error otherwise.